+++ controllers/index.js
@@ -1,9 +1,9 @@
 const router = require('express').Router();
-
+const homeRoutes = require('./homeRoutes')
 const apiRoutes = require('./api');
 
 router.use('/api', apiRoutes);
-
+router.use('/', homeRoutes)
 router.use((req, res) => {
   res.status(404).end();
 });
